Fixed  check and mate detection
finished V1,
This means that the system is able to determine the difference between a check and a check mate, 

It checks if the king is under check, if he is, then the king will automatically check for mate, it does this by figuring out which squares it can possibly move to, then preforms the check operation on each of those squares. 

The system has been massively overhauled to know the difference between a n empty space, and a space that is not there.

The mate detection does not yet account for pieces that could be moved to block the attack, as that is planned for V2